staking test improvements, stake.sol refactor
